style the modal wrapper
