+++ packages/libs/react-ui/src/components/Modal/Modal.css.ts
@@ -1,4 +1,4 @@
-import { ColorType, sprinkles, vars } from '../../styles';
+import { vars } from '../../styles';
 
 import { style } from '@vanilla-extract/css';
