+++ packages/libs/react-ui/src/components/Modal/Modal.stories.tsx
@@ -1,11 +1,9 @@
-import { Button } from '../Button/Button';
-import { Card } from '../Card/Card';
-
-import { IModalProps, Modal } from './Modal';
-import { ModalProvider, useModal } from './ModalProvider';
+import { IModalProps } from './Modal';
+import { ModalProvider } from './ModalProvider';
+import { Content } from './StoryComponents';
 
 import type { Meta, StoryObj } from '@storybook/react';
-import React, { useEffect } from 'react';
+import React from 'react';
 
 const meta: Meta<IModalProps> = {
   title: 'Layout/Modal',
@@ -14,39 +12,6 @@ const meta: Meta<IModalProps> = {
 
 export default meta;
 type Story = StoryObj<IModalProps>;
-
-export const ModalContent = () => {
-  return (
-    <>
-      <Card fullWidth={true}>
-        <h4>Getting Started is Simple</h4>
-        <div>
-          Learn Kadena&apos;s core concepts & tools for development in 15
-          minutes
-        </div>
-      </Card>
-      <Card fullWidth={true}>
-        <h4>Getting Started is Simple</h4>
-        <div>
-          Learn Kadena&apos;s core concepts & tools for development in 15
-          minutes
-        </div>
-      </Card>
-    </>
-  );
-};
-
-export const Content = () => {
-  const { renderModal } = useModal();
-
-  return (
-    <>
-      <Button onClick={() => renderModal(<ModalContent />)} title="openModal">
-        Open modal
-      </Button>
-    </>
-  );
-};
 
 export const Primary: Story = {
   name: 'Modal',
